requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor
beanie
pydantic>=2.0.0
pydantic[email]
python-multipart
python-jose
cryptography
passlib[bcrypt]
python-dotenv
bcrypt



